requirements.txt
Flask~=2.3.2
PyChromecast~=14.0.5
gitpython~=3.1.32
js2py~=0.74
MarkupSafe~=2.1.3
ffmpeg-python~=0.2.0
Werkzeug~=2.3.6
pathvalidate~=3.2.0
requests~=2.32.3
Flask-Minify~=0.50
# For development and testing
pytest~=8.4.1
pytest-cov~=6.2.1
pytest-mock~=3.14.1
